Separation of demo code into modules, to learn how Go modules work
